Rake tasks cookbook:test and cookbook:dev_test can be run against a chef repo.
Case 1: The cwd is a chef repo, strainer will be run on each cookbook in the folder cookbooks.
Case 2: The cwd is a cookbook repo, strainer will be run on this cookbook.
Case 3: The user uses the option "name" e.g "name=cookbook1,cookbook2", so strainer will be run on each cookbook mentioned in the option "name" only.